landing_page_generator/pyproject.toml
[tool.poetry]
name = "landing_page_generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
python-dotenv = "^1.0.1"
crewai = {extras = ["tools"], version = "^0.63.6"}

[tool.poetry.scripts]
landing_page_generator = "landing_page_generator.main:run"
train = "landing_page_generator.main:train"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"